Moved main() from index.py to report.py
